+++ Silkrode/app/src/main/res/layout/fragment_play.xml
@@ -17,7 +17,9 @@
         <com.google.android.material.appbar.MaterialToolbar
             android:id="@+id/toolbar"
             android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize">
+            android:layout_height="?attr/actionBarSize"
+            app:navigationIcon="@drawable/abc_vector_test"
+            tools:ignore="PrivateResource">
 
             <ProgressBar
                 android:id="@+id/progress_bar_loading"
